\name{uhecrauger2014}
\docType{data}
\alias{uhecrauger2014}
\title{Auger 2014 Ultra-high Energy Cosmic Ray Events}
\description{
  This data set gives Equatorial coordinates and energy of the highest energy events observed by the Pierre Auger Observatory up to 2014.
}
\usage{uhecrauger2014}
\format{A data frame with three columns. The first column contains the right ascension
 of the arrival diretion, the second the declination and the third the cosmic ray energy.}
\source{\href{https://www.auger.org/index.php/document-centre/finish/115-data/2354-list-of-ultra-high-energy-cosmic-ray-events-2014}{Pierre Auger Observatory 2014 event list}}
\references{
  A. Aab et al. (The Pierre Auger Collaboration), 2014, "Searches for Anisotropies in
  the Arrival Directions of the Highest Energy Cosmic Rays Detected by the       
  Pierre Auger Observatory", \href{https://arxiv.org/abs/1411.6111}{arXiv:1411.6111}.
}
\keyword{datasets}